[test] Run qiskit remote tests in CI
